LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-report-only-expected.txt

CONSOLE MESSAGE: The Content Security Policy directive 'sandbox' is ignored when delivered in a report-only policy.
CONSOLE MESSAGE: The Content Security Policy 'sandbox' was delivered in report-only mode, but does not specify a 'report-uri'; the policy will have no effect. Please either add a 'report-uri' directive, or deliver the policy via the 'Content-Security-Policy' header.
ALERT: Script executed in iframe.
ALERT: PASS: Iframe was not in a unique origin
